requirements.txt
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
websocket-client==1.6.4
wsproto==1.2.0
